fix lint.

+++ src/c_api_common.cc
@@ -25,7 +25,7 @@ DLManagedTensor* CreateTmpDLManagedTensor(const DGLArgValue& arg) {
 PackedFunc ConvertNDArrayVectorToPackedFunc(const std::vector<NDArray>& vec) {
     auto body = [vec](DGLArgs args, DGLRetValue* rv) {
         const int which = args[0];
-        if (which >= vec.size()) {
+        if (which >= static_cast<int>(vec.size())) {
             LOG(FATAL) << "invalid choice";
         } else {
             *rv = std::move(vec[which]);
